# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Inner Circle Partners Portal API",
    "description": "API documentation for the Inner Circle Partners Portal - a comprehensive partner referral management system.\n\n## Authentication\n\nMost endpoints require authentication via Clerk. Include the session token in your requests.\n\n## Rate Limiting\n\nAPI requests are rate-limited to prevent abuse:\n- General endpoints: 100 requests per minute\n- Auth endpoints: 10 requests per minute\n- Webhook endpoints: 1000 requests per minute\n\n## Response Format\n\nAll responses follow a standard format:\n```json\n{\n  \"success\": true,\n  \"data\": { ... },\n  \"meta\": { \"timestamp\": \"...\" }\n}\n```\n\nError responses:\n```json\n{\n  \"success\": false,\n  \"error\": {\n    \"code\": \"ERROR_CODE\",\n    \"message\": \"Human readable message\"\n  }\n}\n```",
    "version": "1.0.0",
    "contact": {
      "name": "Inner Circle Support",
      "email": "support@innercircle.co"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://innercircle.co/terms"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.innercircle.co",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/partners/{id}/analytics": {
      "get": {
        "tags": ["Partners"],
        "summary": "Get partner analytics",
        "description": "Returns time-series analytics data for charts and reports",
        "operationId": "getPartnerAnalytics",
        "security": [{ "clerkAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "period",
            "in": "query",
            "description": "Time period for analytics",
            "schema": {
              "type": "string",
              "enum": ["7d", "30d", "90d", "365d"],
              "default": "30d"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsResponse"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/partners/{id}/payouts": {
      "get": {
        "tags": ["Payments"],
        "summary": "List partner payouts",
        "description": "Returns payout history for a partner",
        "operationId": "listPayouts",
        "security": [{ "clerkAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["PENDING", "PROCESSING", "COMPLETED", "FAILED", "CANCELLED"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of payouts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutsListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/partners/{id}/milestones": {
      "get": {
        "tags": ["Partners"],
        "summary": "Get partner milestones",
        "description": "Returns milestone progress for gamification",
        "operationId": "getMilestones",
        "security": [{ "clerkAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Milestone progress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestonesResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "clerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk session token"
      },
      "adminRole": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Role",
        "description": "Requires admin role in Clerk user metadata"
      }
    },
    "schemas": {
...
      "AnalyticsResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "earnings": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "date": { "type": "string", "format": "date" },
                    "amount": { "type": "integer" }
                  }
                }
              },
              "conversions": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "date": { "type": "string" },
                    "count": { "type": "integer" }
                  }
                }
              },
              "sources": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "source": { "type": "string" },
                    "clicks": { "type": "integer" },
                    "conversions": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
...
      "PayoutsListResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "payouts": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Payout" }
              },
              "total": { "type": "integer" }
            }
          }
        }
      },
      "Payout": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "amount": { "type": "integer", "description": "Amount in cents" },
          "fee": { "type": "integer" },
          "netAmount": { "type": "integer" },
          "status": { "type": "string", "enum": ["PENDING", "PROCESSING", "COMPLETED", "FAILED", "CANCELLED"] },
          "requestedAt": { "type": "string", "format": "date-time" },
          "processedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "MilestonesResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "milestones": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "name": { "type": "string" },
                    "description": { "type": "string" },
                    "target": { "type": "integer" },
                    "current": { "type": "integer" },
                    "completed": { "type": "boolean" },
                    "reward": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "object" }
            }
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required"
              }
            }
          }
        }
      },
...
    }
  }
}