# src/app/api/partners/[id]/payouts/route.ts
/**
 * Partner Payouts API Endpoint
 * GET /api/partners/[id]/payouts
 * 
 * Returns payout history with:
 * - Upcoming payouts
 * - Past payouts
 * - Breakdown details
 * 
 * Security: Requires authentication (placeholder)
 */

import { NextRequest, NextResponse } from 'next/server';
import { z } from 'zod';
import type { ApiResponse } from '@/types';
import { successResponse, errorResponse, ErrorCodes, generateRandomString } from '@/lib/utils';

// ============================================
// Types
// ============================================

type PayoutStatus = 'PENDING' | 'PROCESSING' | 'COMPLETED' | 'FAILED';
type PayoutMethod = 'BANK_TRANSFER' | 'PAYPAL' | 'CHECK';

interface PayoutBreakdown {
  referrals: number;
  conversionAmount: number;
  bonuses: number;
  adjustments: number;
}

interface Payout {
  id: string;
  amount: number;
  status: PayoutStatus;
  periodStart: Date;
  periodEnd: Date;
  scheduledDate?: Date;
  processedDate?: Date;
  breakdown: PayoutBreakdown;
  method: PayoutMethod;
  transactionId?: string;
}

// ============================================
// Validation
// ============================================

const paramsSchema = z.object({
  id: z.string().min(1, 'Partner ID is required'),
});

const querySchema = z.object({
  status: z.enum(['PENDING', 'PROCESSING', 'COMPLETED', 'FAILED']).optional(),
  limit: z.coerce.number().int().min(1).max(50).optional().default(12),
  offset: z.coerce.number().int().min(0).optional().default(0),
});

// ============================================
// Mock Data Generator
// ============================================

function generateMockPayouts(_partnerId: string): Payout[] {
  const now = new Date();
  
  const payouts: Payout[] = [
    // Upcoming payout
    {
      id: `payout_${generateRandomString(8)}`,
      amount: 124500,
      status: 'PENDING',
      periodStart: new Date(now.getFullYear(), now.getMonth(), 1),
      periodEnd: new Date(now.getFullYear(), now.getMonth() + 1, 0),
      scheduledDate: new Date(now.getFullYear(), now.getMonth() + 1, 5),
      breakdown: {
        referrals: 47,
        conversionAmount: 118500,
        bonuses: 10000,
        adjustments: -4000,
      },
      method: 'BANK_TRANSFER',
    },
  ];

  // Generate past 6 months of payouts
  for (let i = 1; i <= 6; i++) {
    const monthStart = new Date(now.getFullYear(), now.getMonth() - i, 1);
    const monthEnd = new Date(now.getFullYear(), now.getMonth() - i + 1, 0);
    const scheduledDate = new Date(now.getFullYear(), now.getMonth() - i + 1, 5);
    const processedDate = new Date(scheduledDate.getTime() + Math.random() * 24 * 60 * 60 * 1000);

    const referrals = Math.floor(25 + Math.random() * 50);
    const conversionAmount = referrals * (2000 + Math.floor(Math.random() * 1500));
    const bonuses = Math.random() > 0.5 ? Math.floor(Math.random() * 15000) : 0;
    const adjustments = Math.random() > 0.8 ? -Math.floor(Math.random() * 5000) : 0;

    payouts.push({
      id: `payout_${generateRandomString(8)}`,
      amount: conversionAmount + bonuses + adjustments,
      status: 'COMPLETED',
      periodStart: monthStart,
      periodEnd: monthEnd,
      scheduledDate,
      processedDate,
      breakdown: {
        referrals,
        conversionAmount,
        bonuses,
        adjustments,
      },
      method: Math.random() > 0.2 ? 'BANK_TRANSFER' : 'PAYPAL',
      transactionId: `TXN-${monthStart.getFullYear()}-${String(Math.floor(Math.random() * 1000000)).padStart(6, '0')}`,
    });
  }

  return payouts;
}

// ============================================
// Handler
// ============================================

export async function GET(
  request: NextRequest,
  context: { params: Promise<{ id: string }> }
): Promise<NextResponse<ApiResponse<Payout[]>>> {
  try {
    // Await and validate route params
    const params = await context.params;
    const paramsResult = paramsSchema.safeParse(params);

    if (!paramsResult.success) {
      return NextResponse.json(
        errorResponse(
          ErrorCodes.VALIDATION_ERROR,
          'Invalid partner ID',
          { errors: paramsResult.error.flatten().fieldErrors }
        ),
        { status: 400 }
      );
    }

    const { id: partnerId } = paramsResult.data;

    // Parse query parameters
    const { searchParams } = new URL(request.url);
    const queryResult = querySchema.safeParse({
      status: searchParams.get('status') ?? undefined,
      limit: searchParams.get('limit') ?? undefined,
      offset: searchParams.get('offset') ?? undefined,
    });

    if (!queryResult.success) {
      return NextResponse.json(
        errorResponse(
          ErrorCodes.VALIDATION_ERROR,
          'Invalid query parameters',
          { errors: queryResult.error.flatten().fieldErrors }
        ),
        { status: 400 }
      );
    }

    const { status: statusFilter, limit, offset } = queryResult.data;

    // TODO: Authenticate request
    // const authResult = await authenticateRequest(request);
    // if (!authResult || authResult.partnerId !== partnerId) {
    //   return NextResponse.json(
    //     errorResponse(ErrorCodes.UNAUTHORIZED, 'Unauthorized'),
    //     { status: 401 }
    //   );
    // }

    // TODO: Query database
    // const payouts = await prisma.payout.findMany({
    //   where: {
    //     partnerId,
    //     ...(statusFilter && { status: statusFilter }),
    //   },
    //   orderBy: { periodStart: 'desc' },
    //   skip: offset,
    //   take: limit,
    // });

    // For now, generate mock data
    let payouts = generateMockPayouts(partnerId);

    // Apply filters
    if (statusFilter) {
      payouts = payouts.filter(p => p.status === statusFilter);
    }

    // Apply pagination
    payouts = payouts.slice(offset, offset + limit);

    return NextResponse.json(
      successResponse(payouts),
      { status: 200 }
    );
  } catch (error) {
    console.error('GET /api/partners/[id]/payouts error:', error);

    return NextResponse.json(
      errorResponse(
        ErrorCodes.INTERNAL_ERROR,
        'An unexpected error occurred. Please try again.'
      ),
      { status: 500 }
    );
  }
}